moved sdl functions to new file
